Move files to sub-package xof
