+++ library/xof/api/xof.api
@@ -1,16 +1,16 @@
-public abstract class org/kotlincrypto/core/Xof : org/kotlincrypto/core/Algorithm, org/kotlincrypto/core/Copyable, org/kotlincrypto/core/Resettable, org/kotlincrypto/core/Updatable {
-	protected abstract fun newReader ()Lorg/kotlincrypto/core/Xof$Reader;
-	public final fun reader ()Lorg/kotlincrypto/core/Xof$Reader;
-	public final fun reader (Z)Lorg/kotlincrypto/core/Xof$Reader;
-	public static synthetic fun reader$default (Lorg/kotlincrypto/core/Xof;ZILjava/lang/Object;)Lorg/kotlincrypto/core/Xof$Reader;
+public abstract class org/kotlincrypto/core/xof/Xof : org/kotlincrypto/core/Algorithm, org/kotlincrypto/core/Copyable, org/kotlincrypto/core/Resettable, org/kotlincrypto/core/Updatable {
+	protected abstract fun newReader ()Lorg/kotlincrypto/core/xof/Xof$Reader;
+	public final fun reader ()Lorg/kotlincrypto/core/xof/Xof$Reader;
+	public final fun reader (Z)Lorg/kotlincrypto/core/xof/Xof$Reader;
+	public static synthetic fun reader$default (Lorg/kotlincrypto/core/xof/Xof;ZILjava/lang/Object;)Lorg/kotlincrypto/core/xof/Xof$Reader;
 	public final fun toString ()Ljava/lang/String;
 	public final fun use (Lkotlin/jvm/functions/Function1;)Ljava/lang/Object;
 	public final fun use (ZLkotlin/jvm/functions/Function1;)Ljava/lang/Object;
-	public static synthetic fun use$default (Lorg/kotlincrypto/core/Xof;ZLkotlin/jvm/functions/Function1;ILjava/lang/Object;)Ljava/lang/Object;
+	public static synthetic fun use$default (Lorg/kotlincrypto/core/xof/Xof;ZLkotlin/jvm/functions/Function1;ILjava/lang/Object;)Ljava/lang/Object;
 }
 
-public abstract class org/kotlincrypto/core/Xof$Reader {
-	public fun <init> (Lorg/kotlincrypto/core/Xof;)V
+public abstract class org/kotlincrypto/core/xof/Xof$Reader {
+	public fun <init> (Lorg/kotlincrypto/core/xof/Xof;)V
 	public final fun bytesRead ()J
 	public final fun close ()V
 	protected abstract fun closeProtected ()V
@@ -22,20 +22,20 @@ public abstract class org/kotlincrypto/core/Xof$Reader {
 	public final fun use (Lkotlin/jvm/functions/Function1;)Ljava/lang/Object;
 }
 
-public abstract interface class org/kotlincrypto/core/XofAlgorithm : org/kotlincrypto/core/Algorithm {
+public abstract interface class org/kotlincrypto/core/xof/XofAlgorithm : org/kotlincrypto/core/Algorithm {
 }
 
-public abstract class org/kotlincrypto/core/XofFactory {
+public abstract class org/kotlincrypto/core/xof/XofFactory {
 }
 
-protected abstract class org/kotlincrypto/core/XofFactory$XofDelegate : org/kotlincrypto/core/Xof, org/kotlincrypto/core/Algorithm, org/kotlincrypto/core/Resettable, org/kotlincrypto/core/Updatable {
-	protected fun <init> (Lorg/kotlincrypto/core/XofFactory;Lorg/kotlincrypto/core/XofAlgorithm;)V
+protected abstract class org/kotlincrypto/core/xof/XofFactory$XofDelegate : org/kotlincrypto/core/xof/Xof, org/kotlincrypto/core/Algorithm, org/kotlincrypto/core/Resettable, org/kotlincrypto/core/Updatable {
+	protected fun <init> (Lorg/kotlincrypto/core/xof/XofFactory;Lorg/kotlincrypto/core/xof/XofAlgorithm;)V
 	public fun algorithm ()Ljava/lang/String;
 	public final fun equals (Ljava/lang/Object;)Z
-	protected final fun getDelegate ()Lorg/kotlincrypto/core/XofAlgorithm;
+	protected final fun getDelegate ()Lorg/kotlincrypto/core/xof/XofAlgorithm;
 	public final fun hashCode ()I
-	protected final fun newReader ()Lorg/kotlincrypto/core/Xof$Reader;
-	protected abstract fun newReader (Lorg/kotlincrypto/core/XofAlgorithm;)Lorg/kotlincrypto/core/Xof$Reader;
+	protected final fun newReader ()Lorg/kotlincrypto/core/xof/Xof$Reader;
+	protected abstract fun newReader (Lorg/kotlincrypto/core/xof/XofAlgorithm;)Lorg/kotlincrypto/core/xof/Xof$Reader;
 	public fun reset ()V
 	public fun update (B)V
 	public fun update ([B)V
