+++ library/xof/src/commonMain/kotlin/org/kotlincrypto/core/xof/Xof.kt
@@ -13,8 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  **/
-package org.kotlincrypto.core
+package org.kotlincrypto.core.xof
 
+import org.kotlincrypto.core.*
 import org.kotlincrypto.endians.BigEndian.Companion.toBigEndian
 import kotlin.jvm.JvmName
 import kotlin.jvm.JvmOverloads
