+++ library/xof/src/commonMain/kotlin/org/kotlincrypto/core/xof/XofAlgorithm.kt
@@ -13,7 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  **/
-package org.kotlincrypto.core
+package org.kotlincrypto.core.xof
+
+import org.kotlincrypto.core.Algorithm
 
 /**
  * Denotes a class as a user of a specified cryptographic [algorithm]
