+++ library/xof/src/commonMain/kotlin/org/kotlincrypto/core/xof/XofFactory.kt
@@ -13,7 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  **/
-package org.kotlincrypto.core
+package org.kotlincrypto.core.xof
+
+import org.kotlincrypto.core.*
 
 /**
  * Factory class for implementors of Digest or Mac to wrap them
