+++ library/xof/src/commonTest/kotlin/org/kotlincrypto/core/xof/XofUtilsUnitTest.kt
@@ -13,8 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  **/
-package org.kotlincrypto.core
+package org.kotlincrypto.core.xof
 
+import org.kotlincrypto.core.InternalKotlinCryptoApi
 import kotlin.test.Test
 import kotlin.test.assertContentEquals
